// src/types/financial.ts
export interface EfficiencyRatios {
  assetTurnover: number;
  fixedAssetTurnover: number;
  inventoryTurnover: number;
  dio: number;
  arTurnover: number;
  dso: number;
  apTurnover: number;
  dpo: number;
  ccc: number;
}